+++ src/main/java/kitchenpos/order/domain/TableGroup.java
@@ -1,14 +1,12 @@
-package kitchenpos.table.domain;
+package kitchenpos.order.domain;
 
 import java.time.LocalDateTime;
-import java.util.List;
 import java.util.Objects;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import org.springframework.util.CollectionUtils;
 
 @Entity
 public class TableGroup {
