+++ src/main/java/kitchenpos/order/domain/repository/OrderTableRepository.java
@@ -1,7 +1,7 @@
-package kitchenpos.table.domain.repository;
+package kitchenpos.order.domain.repository;
 
 import java.util.List;
-import kitchenpos.table.domain.OrderTable;
+import kitchenpos.order.domain.OrderTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
