+++ src/main/java/kitchenpos/order/service/OrderService.java
@@ -13,8 +13,8 @@ import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.order.service.dto.OrderRequest;
 import kitchenpos.order.service.dto.OrderResponse;
 import kitchenpos.order.service.dto.OrderStatusRequest;
-import kitchenpos.table.domain.OrderTable;
-import kitchenpos.table.domain.repository.OrderTableRepository;
+import kitchenpos.order.domain.OrderTable;
+import kitchenpos.order.domain.repository.OrderTableRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
