+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -13,7 +13,7 @@ import kitchenpos.menugroup.service.dto.MenuGroupResponse;
 import kitchenpos.menu.service.dto.MenuResponse;
 import kitchenpos.order.service.dto.OrderResponse;
 import kitchenpos.product.service.dto.ProductResponse;
-import kitchenpos.table.service.dto.TableResponse;
+import kitchenpos.order.service.dto.TableResponse;
 import kitchenpos.menu.service.MenuService;
 import kitchenpos.menugroup.service.MenuGroupService;
 import kitchenpos.order.service.OrderService;
@@ -24,7 +24,7 @@ import kitchenpos.supports.MenuGroupFixture;
 import kitchenpos.supports.OrderFixture;
 import kitchenpos.supports.OrderTableFixture;
 import kitchenpos.supports.ProductFixture;
-import kitchenpos.table.service.TableService;
+import kitchenpos.order.service.TableService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
