+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -6,13 +6,13 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
 import kitchenpos.order.service.dto.OrderStatusRequest;
-import kitchenpos.table.service.dto.TableGroupRequest;
+import kitchenpos.order.service.dto.TableGroupRequest;
 import kitchenpos.menugroup.service.dto.MenuGroupResponse;
 import kitchenpos.menu.service.dto.MenuResponse;
 import kitchenpos.order.service.dto.OrderResponse;
 import kitchenpos.product.service.dto.ProductResponse;
-import kitchenpos.table.service.dto.TableGroupResponse;
-import kitchenpos.table.service.dto.TableResponse;
+import kitchenpos.order.service.dto.TableGroupResponse;
+import kitchenpos.order.service.dto.TableResponse;
 import kitchenpos.menu.service.MenuService;
 import kitchenpos.menugroup.service.MenuGroupService;
 import kitchenpos.order.service.OrderService;
@@ -24,8 +24,8 @@ import kitchenpos.supports.OrderFixture;
 import kitchenpos.supports.OrderTableFixture;
 import kitchenpos.supports.ProductFixture;
 import kitchenpos.supports.TableGroupFixture;
-import kitchenpos.table.service.TableService;
-import kitchenpos.table.service.TableGroupService;
+import kitchenpos.order.service.TableService;
+import kitchenpos.order.service.TableGroupService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
